Split test command declared env tests
